--- sdk/monitor/azure-opentelemetry-exporter-azuremonitor/azure/opentelemetry/__init__.py
@@ -1,1 +1,0 @@
-__path__ = __import__('pkgutil').extend_path(__path__, __name__)
