+++ sdk/monitor/azure-opentelemetry-exporter-azuremonitor/azure/opentelemetry/sdk/auto_collection/__init__.py
@@ -1,6 +1,6 @@
 # Copyright (c) Microsoft Corporation. All rights reserved.
 # Licensed under the MIT License. See License.txt in the project root for license information.
-from azure.opentelemetry.sdk.azuremonitor.auto_collection._auto_collection import (
+from azure.opentelemetry.sdk.auto_collection._auto_collection import (
     AutoCollection,
     standard_metrics_processor
 )
