+++ sdk/monitor/azure-opentelemetry-exporter-azuremonitor/azure/opentelemetry/sdk/auto_collection/_utils.py
