+++ sdk/monitor/azure-opentelemetry-exporter-azuremonitor/tests/auto_collection/test_performance_metrics.py
@@ -8,8 +8,8 @@ from unittest import mock
 from opentelemetry import metrics
 from opentelemetry.sdk.metrics import MeterProvider, Observer
 
-from azure.opentelemetry.sdk.azuremonitor.auto_collection._performance_metrics import PerformanceMetrics
-from azure.opentelemetry.sdk.azuremonitor.auto_collection._utils import AutoCollectionType
+from azure.opentelemetry.sdk.auto_collection._performance_metrics import PerformanceMetrics
+from azure.opentelemetry.sdk.auto_collection._utils import AutoCollectionType
 
 
 def throw(exc_type, *args, **kwargs):
@@ -170,10 +170,10 @@ class TestPerformanceMetrics(unittest.TestCase):
             obs.aggregators[tuple(self._test_labels.items())].current, 50
         )
 
-    @mock.patch("azure.opentelemetry.sdk.azuremonitor.auto_collection._performance_metrics.psutil")
+    @mock.patch("azure.opentelemetry.sdk.auto_collection._performance_metrics.psutil")
     def test_track_process_cpu(self, psutil_mock):
         with mock.patch(
-            "azure.opentelemetry.sdk.azuremonitor.auto_collection._performance_metrics.PROCESS"
+            "azure.opentelemetry.sdk.auto_collection._performance_metrics.PROCESS"
         ) as process_mock:
             performance_metrics_collector = PerformanceMetrics(
                 meter=self._meter,
@@ -195,10 +195,10 @@ class TestPerformanceMetrics(unittest.TestCase):
                 obs.aggregators[tuple(self._test_labels.items())].current, 22.2
             )
 
-    @mock.patch("azure.opentelemetry.sdk.azuremonitor.auto_collection._performance_metrics.logger")
+    @mock.patch("azure.opentelemetry.sdk.auto_collection._performance_metrics.logger")
     def test_track_process_cpu_exception(self, logger_mock):
         with mock.patch(
-            "azure.opentelemetry.sdk.azuremonitor.auto_collection._performance_metrics.psutil"
+            "azure.opentelemetry.sdk.auto_collection._performance_metrics.psutil"
         ) as psutil_mock:
             performance_metrics_collector = PerformanceMetrics(
                 meter=self._meter,
@@ -219,7 +219,7 @@ class TestPerformanceMetrics(unittest.TestCase):
 
     def test_track_process_memory(self):
         with mock.patch(
-            "azure.opentelemetry.sdk.azuremonitor.auto_collection._performance_metrics.PROCESS"
+            "azure.opentelemetry.sdk.auto_collection._performance_metrics.PROCESS"
         ) as process_mock:
             performance_metrics_collector = PerformanceMetrics(
                 meter=self._meter,
@@ -242,10 +242,10 @@ class TestPerformanceMetrics(unittest.TestCase):
                 obs.aggregators[tuple(self._test_labels.items())].current, 100
             )
 
-    @mock.patch("azure.opentelemetry.sdk.azuremonitor.auto_collection._performance_metrics.logger")
+    @mock.patch("azure.opentelemetry.sdk.auto_collection._performance_metrics.logger")
     def test_track_process_memory_exception(self, logger_mock):
         with mock.patch(
-            "azure.opentelemetry.sdk.azuremonitor.auto_collection._performance_metrics.PROCESS",
+            "azure.opentelemetry.sdk.auto_collection._performance_metrics.PROCESS",
             throw(Exception),
         ):
             performance_metrics_collector = PerformanceMetrics(
